Update integration tests
Now we always show the index offence page, we must update the
tests to take account of the offence page.

I’ve updated the test that originally tested the index offence page
to take account of the presence of multiple offences and ensure
the application is created with the correct offence.