Add function to log to Sentry
We were seeing an error in Cypress 'Module not found: Can't resolve
'async_hooks'' as Cypress is trying to import the node version of the
package in the client. Webpack doesn't seem to bundle the correct
version internally in Cypress for some reason. So we have to
conditionally import the Sentry function in the correct environment.
Hopefully this will be fixed in a future version of Cypress