Allow findSummary function to continue if no summary is found
We have seen an error in prod where a Summary label could not be found
which caused the application to throw a journey-ending error.
As the label returned from this function is shown in the CYA screen but
*not* during the application itself we have deemed it acceptable to show
an empty string instead of the label. We are hoping this will be very
rare but of course we need to know when this happens so we pass to the
exception to Sentry.