Test behaviour of actionsCell at lower level
Previously we were testing the return value of the actionsCell by
manipulating the input to the applicationTableRow which involved some
uneccessary setup and obscured the behaviour under test. By testing at
the appropriate level we require less code and it's easier to work out
what's going on.

I've also correct the description of one of the tests