Update Withdrawals tests
These tests relied on the user being able to see the 'Withdraw' link on
the application dashboard when the application was in any state. Now
that only pre-submitted ('started' status) applications are shown on the
dashboard we can't start the journey there so instead we'll navigate
directly to the Withdrawable select page instead, except for
applications.