Add new function to map Requests For Placement to Summary Cards
We need to map the RfPs which are returned from the new endpoint to GDS
Summary Cards. Request For Placements don't have many of the fields that
we used to use for displaying Placement Applications, we show the data
that we did have at the top level by rendering out the RfPs `document`
key. We also add a status tag so users know where the RfP is up to.
Following the lead taken in submittedApplicationSummaryCards we use an
OOP approach here.