Await result of retrieveFlags
Previously we were awaiting the result of the previous call to the flag
service before beginning the next call. This meant that the final call
was not being awaited.
I have switched out the forEach for a map and added a Promise.all
meaning that all the values for the feature switches will be populated
before the function exits.