Remove docker container
#1793 broke our CI pipeline by introducing a higher version of
@playwright/test to our package.json than was available in our CI
pipeline. In CI we used the Microsoft Docker Image which comes with the
dependencies that Playwright needs preinstalled, this container supports
one version of Playwright specified in the image name. However the
version of Playwright that is run is specified in the package.json.
So when Renovate helpfully upgraded Playwright to the latest version in
the package.json it didn't match the version that was actually being
used (the docker container) which caused an error.
#1799 ensured that we used the same version of Playwright in the
package.json as a temporary fix but now this PR ensures that rather than
use the Docker container prebuilt with Playwright we use npx to install
playwright which will mean the version used will be compatible. This may
mean it takes a little while longer for the E2Es to run but I believe
the pay off is worth it in reduced maintenance burden.