Iterate on Additional Documents page reponse
When rendering a Request For Placement's `document` we show the result
from the response method. I noticed that the AttachDocuments' response
just showed the file names and description but they don't show what the
files actually were. No we add a description to say 'Additional files'
before rendering the file names and descriptions.
As a further iteration we should probably include a link to the file at
this point so they user can access the file but this is out of scope for
now