Migrate Sentry from 7 to 8
Following the guide here:
https://docs.sentry.io/platforms/javascript/guides/express/migration/v7-to-v8.
A few pertinent highlights:
- Sentry.init must be called first of all the calls that are made in the
application. The docs suggested having Sentry.init in a seperate
'./instrumentation.ts' file, but I don't see the need for this as long
as it called first.
- We no longer need to individually call Sentry.Integrations.Http and
Tracing.Integrations.Express - these are setup automatically in v8.
- We no longer need to pass Sentry.Handlers.requestHandler and
Sentry.Handlers.tracingHandler to Express individually. These are now
done by a single Sentry.SetupExpressErrorHandler function which takes
the Express 'app' as an argument.
- The docs state: 'We now support the following integrations out of the
box with 0 configuration: httpIntegration: Automatically instruments
Node http and https standard libraries...'. I am unsure as to whether
this means that there's no need to call Sentry.httpIntegration in the
integrations property of Sentry.init or whether httpIntegration simply
needs *no configuration now* so for safety I have left the call in.