Don't run E2Es on PR
We did run E2Es on PR in order to prove that the currently deployed code
wasn't failing the E2Es.
The issue with this is that because many PRs can be open these tests can
run *a lot* and seperate test runs end up tripping each other up as they
need a level of stability to be able to work.
We can remove the runs on PR now and monitor if this improves the
general stability of the 'merge' E2Es.