Don't skip to next page on application dashboard to find withdrawn application
#1812 introduced a change to show the most recent applications first by
default.
The E2E tests still assumed that the most recent application was shown
last so would skip forward to the last application dashboard page if
there was more than one.
As there was sometimes only one page which would contain the withdrawn
application this meant that sometimes the tests passed but sometimes
they failed.