Add new Out of Service Bed clients, service and controllers
This is the same as the LostBeds client, service and controller but for
some newly defined 'out of service beds' endpoints.
As our Client tests test against the main branch of the API we will have
to wait until the API side changes are merged before enabling the tests.
The views will be added in a subsequent commit.