Add Out of Service beds UIs
This adds the UIs and view utils for Out of Service beds. They mirror
the Lost Beds views and utils but with all reference to Lost Beds
changed to Out of Service Beds.
These pages are available when the user navigates to the URL directly
but they are currently not available through any links in the
application. That will be added in a subsequent PR.