Move premisesUtils to its own folder
Since initially developing the Manager functionality we have moved from
a naming format of 'nUtils.ts' in the utils folder where 'n' is a
section of functionality or feature to naming a folder after the
functionality or feature and adding the util file(s) there. This allows
for easier mocking and breaking dependency cycles.
In order to be consistent this commit moves the premises utils to that
format. The functionality is unchanged.