Add premisesActions util
Previously the premisesActions were hardcoded into the template but now
we need to be able to configure them by role it makes more sense to have
a util so they are more easily testable.
The premises.cy test suggests that workflow_managers need access to this
functionality, so we will grant them access as well 'manager's (for
backwards compatibility)  too.