Rename bookingActions to legacyBookingActions
In a future commit we will decide which actions to show based on the
role. Rather than add to the complexity present in this function we will
add another function with a name of 'bookingActions' and toggle between
this function (now named 'legacyBookingActions') and the new one. So we
need to rename the existing function to allow us to write a new
'bookingActions' function.