Allow all manager roles to mark bed as Out of Service
This allows all users with a manager role to mark a bed as out of
service in the currently live service (IE not v2).
We need this for the v2Manage/beds.cy.ts integration test to pass.
Otherwise there is an authorization error as the v2Manage E2E test users
don't have the workflow_manager role (and nor should they).
Why change the existing code to make the v2 integration tests pass?
- We believe users are used to their way of working at this point and
users with the 'manager' role will know not to use this button
- the permission will soon be replaced with the 'legacy_manager' role
An alternative would be create a new bedsController in the v2Manage
namespace and copying over the existing beds-related code and removing
this role but we feel that it is too big a piece of work for this PR and
the alternative is a superior choice.