Allow all users to extend & change dates of bookings
As with the previous commit we extend the permissions to all users to
change the dates of bookings.
However unlike the previous commit this should not affect user facing
behaviour as these are controller level permissions. It is highly
unlikely the user will reach the UI to perform this action to begin with
(they would have to use the URL as it is unreachable in the UI).

Permit any user with a manager role to change placement dates

Access to these features will be controls via the UI, which is a
sufficient control during the refactoring period where we move
from v1 Manage to the v2 Manage features in the `/manage/` namespace.