Allow Pact tests to test new namespace
The new namespace's spec is available in a new file
(built-cas1-api-spec.yml not built-api-spec.yml). We need a way to tell
Pact to use this new spec.
As we will gradually move endpoints over to this new namespace and don't
have any plans to introduce any further namespaces I think a simple flag
passed to the `toMatchOpenAPISpec` function will do. We can call
toMatchOpenAPISpec with this flag by using the
'describeCas1NamespaceClient' describe function instead of the current
'describeClient' function.