Use new Cas1OutOfServiceBed entities
This updates the usages of the OutOfServiceBed types to use the new one
defined in the spec and created in the factory we just updated.
There is a lot of changes in this commit but it is mainly just renaming
a handful of fields repeatedly.
We have to handle some cases where nullable properties are undefined in
the tests too.