Use new CAS1 Report Endpoint
This commit downloads reports from the new cas1-specific reports endpoint.

As this new endpoint provides filenames via the Content-Disposition header, this is passed through by the UI. The new filenames included a generated timestamp, so the E2E tests have been updated to reflect this.

This commit also removes unused reports as they are no longer available in the API on the new and old endpoints.