Remove test for navigating to OOS beds from bed list
Now that the entities delivered from the new OOS bed endpoints differ
from the existing OOS bed entities we can no longer use the same views
as before. This is because the views rely on there being a 'startDate'
but the new entities use an 'outOfServiceFrom' date.
We could copy over the controllers and views and update the names of
properties but this is a fair amount of work and a lot of code that we
may not need in future. Instead I think it would be best to readd this
test if we decide we need it at a later date. As a very similar test is
available in the v1 manage folder this should be trivial.