Rename out of service ‘index’ -> ‘premisesIndex’
We now need a separate view for all OOS beds for all APs. This commit renames the existing ‘index’ method, which renders all OOS beds for a single AP, to ‘premisesIndex’, so that we can use ‘index’ for our new view.

It also renames the associated path, API stubs, files etc.