Correct roles for Cancellation test
Previously we used the internal snake_case value but Playwright requires
the value of the Checkbox it is to select in the format of the label the
user will read (in this case 'Title case').
Also we need to give the test user role of 'manager' ('Manage an
Approved Premises (AP)') as the API requires this role to authenticate
the cancellation endpoint.