Run E2Es on PR again
We removed runnning E2Es on PR because we believed they were causing
lots of failures when multiple E2E sets were running concurrently on
seperate PRs.
I think we can re-enable them for now as they will allow us to see if
changes to the E2Es in the PR are valid. We can monitor to see if the
number of failures increases.