Don't run 'soft delete' test locally
This behaviour only exists in CI (and deployed envs). Locally users are
hard deleted. Once they are deleted they cannot be found via search as
they are removed from the DB and only reseeded when the local API is
restarted.In deployed envs the user is soft deleted and can be found via
search.

We may decide to remove this test later or fix the behaviour either
locally or in deployed envs but now is not the time.