Add signIns to Admin tests
These tests require two different roles. This is trivial and clear now
that each tests sign in as the specified user.
I had originally thought that we would need to add an 'afterEach' job to
sign out. This is handled in ./e2e/tests/auth.setup.ts which runs after
every test - so no need for the manual afterEach.