Refactor 'Manage user' spec for clarity
In our spec file we now focus on describing behaviour
and push implementation details into the `editUser` helper.

e.g. our spec now asserts that:

```
editUserPage.shouldHaveAllRolesAssigned(roles)
```

the `editUser` helper is responsible for the DOM related
nitty-gritty:

```
async shouldHaveAllRolesAssigned(labels: ReadonlyArray<Role>) {
  this.assertCheckboxesAreSelected(labels)
}

```

This makes our spec easier to understand and maintain. It describes
the expected *behaviour* from the point of view of the user. In
this case an admin who is managing the roles of another user.