Reintroduce integration test stub
I assume this was removed because it was no longer being used (like the endpoint), but now that we are using the endpoint to retrieve the AP area on the premises page, we reintroduce the single premises stub.

This commit also ensures the API call is stubbed everywhere we are trying to access the premises page in the integration tests.