Remove default sort values and add perPage to OOS client and service
We don’t need default values for sortBy and sortDirection as distinct from what the API defaults to. But we do need to specify a perPage value so that we can ensure that we will render all currently OOS beds for an AP on a single page, by specifying 50 results per page.