Rename OOSB start and end dates to match API
We rename OutOfService date fields:

- outOfServiceFrom -> startDate
- outOfServiceTo -> endDate

to line up with the actual fields implemented at the API, see
[OpenAPI spec][].

---

This commit includes a change to the `errors.json` translation file
which converts field and error keys sent by the API to human error
messages which map to form fields in the DOM -- to which we need to
apply error styling.

The system of "translating" validation errors messages sent from
the API works like this ([see validation error code in the API][]):

1. FROM THE ENTITY LAYER

For validation errors raised from the entity level, by virtue of
a field being non-nullable, the `propertyName` is automatically
used as the field key e.g.

```
data: {
  "title": "Bad Request",
  "status": 400,
  "detail": "There is a problem with your request",
  "invalid-params": [
    {
      "propertyName": "$.endDate",
      "errorType": "empty"
    },
  ]
}
```

2. FROM THE SERVICE LAYER

PROBLEM: HOW TO DISTINGUISH BETWEEN `foo.notes` and `bar.notes`?

Where we have custom validations which are not set at the db level
we can set a bespoke `propertyName` -- but this won't allow us to
have consistent and unique propertyNames e.g. for 3 entities which
all have a `notes` property

 - `LostBedNotes`
 - `OutOfServiceBedNotes`
 - `AnOtherEntityNotes`

[OpenAPI spec]:
https://approved-premises-api-dev.hmpps.service.justice.gov.uk/swagger-ui/index.html?urls.primaryName=CAS1%20API#/out-of-service%20beds/post_premises__premisesId__out_of_service_beds

[validation error code in the API]:
https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/79ed9528719e6829cbac5ecd2a46ce3e21390e72/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/results/ValidatableActionResult.kt#L6-L27