Add timeline tab
This adds some metadata for revisions to the timeline tab.
The revision type could be a list if multiple properties of a revision
have been updated (EG a start date and an end date) so we need to handle
putting a comma in between the non-ultimate list members - I'm suprised
we don't already have a util for this.