Add seperate templates for tabs of OoS bed page
We want to have two seperate tabs with the same navigation. The cleanest
way to do this is to use a partial for each tab. In order to do so we
will need to ensure that the 'activeTab' param has a default value of
'details' so details is the first tab shown when the user lands on the
page.
As the template currently contains the 'bed details' we can move the
markup over to that partial but the timeline one will be empty for now.