Update show OOS bed controller/view
We will render the OOS bed details based on the latest revision on the OOS bed. We also make an additional call to get the bed characteristics to display on the page, and remove the form that previously allowed the user to update the OOS bed details from the same page.