Add description of OoS bed revision
In order to show the description of the revision we need to work out
which properties of the revision have been updated and then which parts
of those properties we want to show to the user. This would be very
unwiedly to do in Nunjucks so it's best done in a TS util function.