Add backlink to OoS bed show page
Previously the Back button just linked to the referrer which would cause
an infite loop if the user visited the 'Timeline' tab. Now we simply
return the user to the OoS bed list page for the 'current' temporality.
The user may have reached the bed by a different temporality so
'current' is a best guess.
Ideally we would implement some breadcrumbs here but we lack the time.