Move 'View OoS beds' test
The next 3 commits will move the new OoS bed tests to a more rational
file structure. They will be stored in folders based on the role
required to perform the action with a file for each area of
functionality. We begin by moving the tests that proves CRU manages can
view all out of service beds.