Extract method to verify bed details
We will need this on multiple Pages soon so to remove duplication I have
extracted this method to the top level Page class.
There's a lot of methods on the Page class and it would make things
clearer if we used more inheritance, but this is out of scope for now.