Overwrite existing bed fields with user input
Before this change if the API threw an error when the page was loaded
the inputs of the form would be populated with the data from the OoS bed
not the erroneously inputted data. We want to show the erroneous data so
the user can see where they've gone wrong and correct their mistake.
I'm not happy with the amount of casting I've had to do here but time
doesn't allow for type assertions currently.