Translate characteristic.propertyName to a positive statement
Previously we were using `Characteristic.name` which is a string
posing a question such as:

> Is this a single room?

or

> Is there a smoke/heat detector in the room?

With this commit we introduce a

```
characteristicUtils.translateCharacteristic(CharacteristicPair)
```

function which attempts to lookup `Characteristic.propertyName` in
a JSON translation file (`i18n/en/characteristics.json`) to find
a label for the characteristic which is cast in the form of a
positive statement form, e.g.

> Single room

or

> Smoke/heat detector

If a translation for the given `propertyName` isn't found we
return the `propertyName`. These are (just about) human readable
e.g.

- `acceptsSexOffenders`
- `acceptsChildSexOffenders`