Don't overwrite date values added to the form
If the user enters an invalid date or doesn't enter a date we want to
present the erroneous input back to them in the form inputs. However by
overwriting the dates in overwriteOoSBedWithUserInput we were causing
any invalid dates to break the form as they weren't able to be parsed by
isoDateToDateInputs.
Meanwhile by passing the spread of userInput into the view we were
already prepopulating the date inputs anyway.