Introduce feature flag into Placement Request actions
When the feature switch is on we will show the new journey to match to a
bed instead of the existing method of placing a person via Placement
Request.
We may decide to retain both journeys in future which should be simple.
I am not sure if the environment variables come through as the values
supplied in the values-[env].yml natively or if they are converted to a
string so we will handle both possibilities in the code.