Fix integration tests
These tests were skipped so were quite out of date. This commit
reinstates them.
Ideally we would start the test at the CRU Dashboard and follow it
through to making a booking and marking a placement as unable to match
but this would require us to flip the feature flag in the Cypress tests
which isn't possible without restarting the server so for now we will
start with the Search page.