Rename bed search to space search
In the new development of the 'Match' transaction users will be
searching for a space to house their PoP as opposed to a bed as was
previously the case. Letting the AP choose the space rather than
specificying a precise bed allows them to better manage their capacity.

These changes are a simple renaming and don't effect any functionality.