Genericise toContainManageAction expectation
This expectation can be really useful with any [Button
Menu](https://design-patterns.service.justice.gov.uk/components/button-menu/)
that takes Actions. I have opted to make the expectation less specific
in order to be able to use it more widely at the cost of the specitivity
of the name. However it should be clear from the usage which actions are
under test.
We will use this expectation later in this PR.