Fix historic out-of-service beds not displaying in dashboard
When using the `/manage/out-of-service-beds` dashboard, historic
out-of-service beds are not displayed when the 'Historic' tab is
selected.

This is because the UI is using the value `'historic'`, which is not a
valid value for the `Temporality` type. The correct value is `'past'`,
and using this value resolves the issue.

No user-facing changes have been made, so the tab still uses the
'Historic' label.