Adapt card title
We no longer want to link to the booking page via the heading. Instead
we want a link to 'View spaces'. The mock had this below the Summary
Card but this is a bit difficult to achieve so instead I've decided to
use the 'actions' property of the Summary card and link in the heading
row instead.