Add match utils for displaying booking details
As well as the above, filterOutAPTypes removes any required characteristics from the placement request that are related to the premises type rather than a room or space. Requirements for a specific type of premises should be captured by the AP type being booked, and will be displayed separately on the booking page.