Update of Placement Request to a Space Search
When the user lands on this page initially the inputs will be
prepopulated with the details from the placement request and the search
will be performed using these parameters.
This is a requirement for later in the development process of the
feature but given this update is required to keep the compiler happy
it's less work to do it now.