Update rendering of filters
Previously there were 3 groups with checkboxes for 'Type of AP',
'Placement requirements' and 'Risk and offences to consider' which all
used the same value for the input ('requiredCharacteristics') and were
sent to the API as one list.
In the new world there is a requirement to send the AP Type seperately
and to add a new group of checkboxes for Gender. This means the
different checkboxes will need different name values which means
reconfiguring this helper function. It would be simpler to have simply
seperate inputs for each checkbox group but given most of this
functionality already exists and we endeavour to keep things DRY I think
it's worth keeping the functionality bundled together.

Because we're using the empty array notation after the input names (EG
apType[]) the results will always be in an array so we don't need the
test to test if a single requirement is sent to the API