Update Space Search results
The mocks call for different information displayed in the results so we
need to update the functions that create the components. There was some
information that isn't relevant to Spaces (IE anything about a bed) that
can be removed. Also we no longer link via the heading but through a
seperate 'View spaces' link