Filter placementCriteria to SpaceCharacteristic
This function filters the AP Types out of an array of PlacementCriteria
which means they will fit the type of a Cas1SpaceCharacteristic. In
order to keep things DRY and to reduce the number of sources of truth we
can use this function. If the SpaceCharacteristics and PlacementCriteria
diverge in future we may find it more useful to maintain seperate lists.