Use factory for body params
The factory produces the same object that the form will produce with the
advantage that if the factory updates we will get those updates in the
controller for free. If the type is updated then we will see compiler
errors and know to update the factory. Both of these factors will
improve the accuracy and correctness of our tests