Add step to click the placement request
We need a function to select a placement request from the CRU dashboard.
Unlike with the assessment there isn't an application ID we can use to
select the PR so instead we will have to select it as the user would,
using the details available on screen that were entered in Apply and
Assess.
For now I am assuming that the Application will be made the same day as
we will be attempting to find it and that it is not parole but these
details can be extracted from Apply/Assess later if needs be