Add step to ensure details from Apply/Asses shown in Match
We want to ensure that search works properly and that it shows the
details entered in the earlier stages of the transaction. We take the
information that was entered and ensure it is visible. We also ensure
the search functions as intended. Currently the filtering doesn't work
but this can be tested here in future.