Add classes to dates on MatchingInformation page
We need an easy way to get this information from the Matching page. We
could do lots of navigating around the DOM but this is relatively
painful in Playwright. We could add test-id's like we do for Cypress
tests but this would require us rewriting
suggestedStaySummaryListOptions to return HTML instead of JS objects.
The alternative is adding classes that Playwright can access easily,
which is what I went for.