Reduce timeout for pagination clicking
I introduced these functions that click through the pages of a paginated
table to find the assessment they need.
Currently the wait the default 5 seconds before clicking to the next
page and searching again. This change reduces this to 1 second which
should speed things up. Locally I have been successful with a timeout of
100ms but I am not sure dev is fast enough for this.
For now I think 1000ms is enough of an increase in speed but if we want
to go even fast we can increase this.