Switch check for click
Check will fail if the item is already clicked and given there is no way
(that I know of) to clear a radio input (once one item is selected one
item will always be selected) this is not desirable. This may be a bug
in Playwright and could be revert at a later date.