Make dates of placement nullable
This may not always exist and we need to handle this case in the tests.
The code for search for the correct row on the CRU dashboard isn't as
eloquent as I would like. We assert the that all the correct cells exist
and then return the final cell's locator meaning we check it exists
twice but I think this is tolerable for now.