Add E2E test for booking a space
We have the test infrastructure in place for searching for a space, now we need it for booking one.

It would be nice to check that the essential and desirable criteria are also present on the space booking page, but the mapping that has taken place between the placement request criteria and the space booking criteria makes that difficult, so I will leave it for future iterations or resolutions of this discrepancy.