Fix space booking E2E tests
We have discovered that the isVisible() assertion doesn’t behave as expected, so we replace these with more reliable expect().toBeVisible() assertions.

I have also removed the test checking the placement requirements are on the space search page. As it stands the requirements are distinct from the characteristics on the space search page, with the former being mapped to the latter. But this is a bit flaky and may change, so for now, we will remove the test and leave the possibility of reinstating it in a future iteration.