Remove ‘manager’ and ‘legacy_manager’ options
These two roles should no longer be used going forward. This commit removes them as options from the manage user page to ensure they are no longer assigned to users. Once this change is in production the roles will be removed from the API and we can then remove all references to them in the UI code.