Use /users/summary to populate dropdown
Previously we used the /users endpoint to populate the user’s dropdown on the tasks page. This endpoint is very heavy weight and returned lots of information when we only needed the user’s id and name.

This commit instead uses the lightweight /user/summary endpoint to populate the dropdown.