Rewrite api paths definitions
This does away with intermediary definitions which make the resulting
object difficult to read and eases jumping to the actual path
definition.

The suffix 'Path' has been removed from all path variables: it is
redundant in this file.

Strategy is to use the plural name of the entity for base paths and
paths that return collections, and use the suffix 'Single' for paths
returning a single entity.