Move date changes functionality fully to v2
This functionality was duplicated between v1 and v2, sharing the same
controller with simply updated path. However, v1 was allowing the
workflow_manager role, while v2 was only allowing the future_manager
role -- we update v2 to allow both roles for now, so it will be easier
to deprecate the workflow_manager role.