Manage E2E: edit keyworker
Adds a new 'manage' set of steps at teh end of the Standard
application/assessment workflow, books a space, logs out the CRU user,
signs in as a Future manager, finds the booking on the premises page,
and assigns a random keyworker to it.

Also ensures the release date is more variable, so specific entries
can be selected more precisely later on.