Add filter form for occupancy view
This change implements the form and its functionality. It removes the
need for passing the start date and duration of the placement in the
link to the occupancy view, as if this is missing it is simply inferred
from the placement request data.

This also ensures the dates shown in the placement request detail
summary are those of the placement request itself, rather than the
filtered view.