Prevent infinite loop for ineligible application
This prevents an infinite loop when building the task list for an
application that has the ESAP AP type selected but no agreement with
CHPP. The loop was the result of the ineligibility page returning the user
to the AP type selection page on submission. This works if the user then saves a
different AP type straight away; however, should the user then exit the
application, any further attempt to load the application would result in
the loop.

To prevent this, the ESAP ineligibility page no longer posts, but simply
offers a link to the AP Type selection.

In addition, to ensure the AP Type task is shown as 'In progress' in the
above scenario, the ESAP ineligibility page always returns an error.
This error is never shown, but prevents the loop from continuing with
processing any further task.