Throw if page already visited
When building the main task list or fetching statuses for building
un-submitted application summaries, some answer combinations may
unintentionally produce an infinite loop due to how the next page to
check is determined.

To prevent this, a simple trail of pages that have already been visited
is built whenever such a loop is running. If a page has already been
visited, an error is now thrown: this prevents the current process
running infinitely, and will alert us of the trail of pages that caused
the infinite loop, so the issue can be remedied.