Fix flaky occupancy test
The test setup may create a mocked placement request which does nto
contain any room-level requirements. When this is the case, no criteria
is added to the placeholder URL used to generate calendar links.

However, the test setup always specified the query string prefix for
this placeholder URL, resulting in intermittent failures.