Fix logic for nameOrPlaceholderCopy
The logic previously assumed the person was a `FullPerson` based solely
on whether the name property was defined. However, this would be false
for an 'unknown' person, but would still render the text for a
'restricted' (LAO) person, which would be incorrect. This updates
`isFullPerson` to check the discriminator instead, which is more
reliable.