Update template rendering of person name
Now uses the new displayName methods for most instances of displaying
the person's name. When the name is displayed in a summary list, it is
now suffixed with ' (Limited access offender)' instead of the prefix
'LAO: ' for better readability.

Some integration tests have been updated to ensure an actual booking is
returned where relevant, rather than a booking summary.