Refactor rendering of lists of characteristics
This combines the similar `requirementsHtml` and
`characteristicsBulletList` methods into one which handles all use
cases.

A better default is returned when the list is empty, showing 'None' in
slightly greyed out text.

For inline characteristics, a default of 'none' is also returned.
Instances of 'Room criteria' show 'no room criteria'.