Cancel previous workflow for the same branch/ref
At the moment, a push to a branch that is currently running the CI
workflow does not cancel it; instead, the workflow continues running in
the background, and results in failure notifications being sent out.
This ensures any previously running workflow for a given branch or ref
is cancelled by newer ones, therefore saving compute and preventing
out-of-date failure notifications being sent out.